clenaup build directories
