+++ src/core/utils/fs.ts
@@ -36,23 +36,3 @@ export async function getPublicFiles(
   const files = await glob('**/*', { cwd: config.publicDir });
   return files.map(unnormalizePath);
 }
-
-/**
- * Recursively remove all directories that are empty/
- */
-export async function removeEmptyDirs(dir: string): Promise<void> {
-  const files = await fs.readdir(dir);
-  for (const file of files) {
-    const filePath = path.join(dir, file);
-    const stats = await fs.stat(filePath);
-    if (stats.isDirectory()) {
-      await removeEmptyDirs(filePath);
-    }
-  }
-
-  try {
-    await fs.rmdir(dir);
-  } catch {
-    // noop on failure - this means the directory was not empty.
-  }
-}
